refact: extrair macro hotbar para module/helpers/macro.mjs
