+++ packs/efeitos-cardigan/CURRENT
@@ -1,1 +1,1 @@
-MANIFEST-001284
+MANIFEST-001286
