--- packs/efeitos-cardigan/LOG
@@ -1,3 +1,0 @@
-2026/06/29-17:13:58.565 377c Recovering log #1282
-2026/06/29-17:13:58.604 377c Delete type=0 #1282
-2026/06/29-17:13:58.604 377c Delete type=3 #1280
