+++ packs/efeitos-cardigan/LOG.old
@@ -1,8 +1,3 @@
-2026/06/29-17:12:10.487 2c24 Recovering log #1278
-2026/06/29-17:12:10.532 2c24 Delete type=0 #1278
-2026/06/29-17:12:10.532 2c24 Delete type=3 #1276
-2026/06/29-17:13:01.218 17d0 Level-0 table #1283: started
-2026/06/29-17:13:01.218 17d0 Level-0 table #1283: 0 bytes OK
-2026/06/29-17:13:01.232 17d0 Delete type=0 #1281
-2026/06/29-17:13:01.265 17d0 Manual compaction at level-0 from '!folders!FszDz3GuPvOHfi8w' @ 72057594037927935 : 1 .. '!items!xY2yZ3zA4bB5cC6D' @ 0 : 0; will stop at (end)
-2026/06/29-17:13:01.265 17d0 Manual compaction at level-1 from '!folders!FszDz3GuPvOHfi8w' @ 72057594037927935 : 1 .. '!items!xY2yZ3zA4bB5cC6D' @ 0 : 0; will stop at (end)
+2026/06/29-17:13:58.565 377c Recovering log #1282
+2026/06/29-17:13:58.604 377c Delete type=0 #1282
+2026/06/29-17:13:58.604 377c Delete type=3 #1280
