+++ packs/equipamentos-cardigan/CURRENT
@@ -1,1 +1,1 @@
-MANIFEST-001295
+MANIFEST-001297
