--- packs/equipamentos-cardigan/LOG
@@ -1,3 +1,0 @@
-2026/06/29-17:13:58.699 377c Recovering log #1293
-2026/06/29-17:13:58.737 377c Delete type=0 #1293
-2026/06/29-17:13:58.737 377c Delete type=3 #1291
