+++ packs/equipamentos-cardigan/LOG.old
@@ -1,8 +1,3 @@
-2026/06/29-17:12:10.830 2c24 Recovering log #1289
-2026/06/29-17:12:10.874 2c24 Delete type=0 #1289
-2026/06/29-17:12:10.874 2c24 Delete type=3 #1287
-2026/06/29-17:13:01.324 17d0 Level-0 table #1294: started
-2026/06/29-17:13:01.324 17d0 Level-0 table #1294: 0 bytes OK
-2026/06/29-17:13:01.341 17d0 Delete type=0 #1292
-2026/06/29-17:13:01.368 17d0 Manual compaction at level-0 from '!folders!A1cE2sS3oR4iO5sF' @ 72057594037927935 : 1 .. '!items!Z4qG7jL1mW5sN9vP' @ 0 : 0; will stop at (end)
-2026/06/29-17:13:01.368 17d0 Manual compaction at level-1 from '!folders!A1cE2sS3oR4iO5sF' @ 72057594037927935 : 1 .. '!items!Z4qG7jL1mW5sN9vP' @ 0 : 0; will stop at (end)
+2026/06/29-17:13:58.699 377c Recovering log #1293
+2026/06/29-17:13:58.737 377c Delete type=0 #1293
+2026/06/29-17:13:58.737 377c Delete type=3 #1291
