+++ packs/racas-cardigan/CURRENT
@@ -1,1 +1,1 @@
-MANIFEST-002243
+MANIFEST-002245
