--- packs/racas-cardigan/LOG
@@ -1,3 +1,0 @@
-2026/06/29-17:13:58.658 27c4 Recovering log #2241
-2026/06/29-17:13:58.694 27c4 Delete type=0 #2241
-2026/06/29-17:13:58.695 27c4 Delete type=3 #2239
