+++ packs/racas-cardigan/LOG.old
@@ -1,8 +1,3 @@
-2026/06/29-17:12:10.616 32dc Recovering log #2237
-2026/06/29-17:12:10.824 32dc Delete type=0 #2237
-2026/06/29-17:12:10.824 32dc Delete type=3 #2235
-2026/06/29-17:13:01.248 17d0 Level-0 table #2242: started
-2026/06/29-17:13:01.248 17d0 Level-0 table #2242: 0 bytes OK
-2026/06/29-17:13:01.265 17d0 Delete type=0 #2240
-2026/06/29-17:13:01.265 17d0 Manual compaction at level-0 from '!items!D6eF9gG2hI5iJ8kL' @ 72057594037927935 : 1 .. '!items!V9wX2yY5zZ8aA1bB' @ 0 : 0; will stop at (end)
-2026/06/29-17:13:01.265 17d0 Manual compaction at level-1 from '!items!D6eF9gG2hI5iJ8kL' @ 72057594037927935 : 1 .. '!items!V9wX2yY5zZ8aA1bB' @ 0 : 0; will stop at (end)
+2026/06/29-17:13:58.658 27c4 Recovering log #2241
+2026/06/29-17:13:58.694 27c4 Delete type=0 #2241
+2026/06/29-17:13:58.695 27c4 Delete type=3 #2239
