+++ packs/skills-cardigan/CURRENT
@@ -1,1 +1,1 @@
-MANIFEST-001312
+MANIFEST-001314
