--- packs/skills-cardigan/LOG
@@ -1,3 +1,0 @@
-2026/06/29-17:13:58.608 377c Recovering log #1310
-2026/06/29-17:13:58.654 377c Delete type=0 #1310
-2026/06/29-17:13:58.654 377c Delete type=3 #1308
