+++ packs/skills-cardigan/LOG.old
@@ -1,8 +1,3 @@
-2026/06/29-17:12:10.537 377c Recovering log #1306
-2026/06/29-17:12:10.609 377c Delete type=0 #1306
-2026/06/29-17:12:10.609 377c Delete type=3 #1304
-2026/06/29-17:13:01.198 17d0 Level-0 table #1311: started
-2026/06/29-17:13:01.198 17d0 Level-0 table #1311: 0 bytes OK
-2026/06/29-17:13:01.218 17d0 Delete type=0 #1309
-2026/06/29-17:13:01.265 17d0 Manual compaction at level-0 from '!folders!QcRacDwarf3GzTk1' @ 72057594037927935 : 1 .. '!items!Y7zZ0aA3bB6cC9dD' @ 0 : 0; will stop at (end)
-2026/06/29-17:13:01.265 17d0 Manual compaction at level-1 from '!folders!QcRacDwarf3GzTk1' @ 72057594037927935 : 1 .. '!items!Y7zZ0aA3bB6cC9dD' @ 0 : 0; will stop at (end)
+2026/06/29-17:13:58.608 377c Recovering log #1310
+2026/06/29-17:13:58.654 377c Delete type=0 #1310
+2026/06/29-17:13:58.654 377c Delete type=3 #1308
